Mon, Feb 7, 2022, 11:29 PM +00:00

+++ CS50x/week2/pset2/substitution_DONE/substitution.c
@@ -88,7 +88,7 @@ string cypher(string ptext)
             c = tolower(c);
             // printf("up is: %i\n", up);
             c %= 97;
-            ptext[i] = key[c];
+            ptext[i] = key[(int) c];
             if (!up)
             {
                 ptext[i] = tolower(ptext[i]);
